pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer"
authors = ["Naumenko_Sergey <sergei.naumencko@yandex.ru>"]
repository = "https://github.com/SergeiNaum/python-project-83"
license = "MIT"
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "3.11.3"
flask = "^2.3.3"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.5"
requests = "^2.31.0"
validators = "^0.21.2"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
